requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.1
django-admin-persian-fonts==0.2
django-celery-beat==2.8.1
django-ckeditor==6.7.2
django-debug-toolbar==5.1.0
django-jalali-date==2.0.0
django-js-asset==3.1.2
django-mptt==0.17.0
django-rosetta==0.10.2
django-taggit==6.1.0
django-timezone-field==7.1
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
kombu==5.5.3
pillow==11.2.1
polib==1.2.0
prompt_toolkit==3.0.51
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==6.1.0
requests==2.32.3
setuptools==79.0.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
dj-database-url==3.0.1
django-redis==6.0.0
psycopg==3.2.12
psycopg-binary==3.2.12
python-dotenv==1.2.1
